requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.7.2
async-timeout==5.0.1
attrs==24.3.0
azure-core==1.32.0
botbuilder-core==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
botframework-streaming==4.16.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
distro==1.9.0
Django==4.2.5
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.8.2
jsonpickle==1.4.2
msal==1.31.1
msrest==0.7.1
multidict==6.1.0
oauthlib==3.2.2
openai==1.58.1
propcache==0.2.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
yarl==1.18.3
